Promote `Digest` from type alias to tuple struct
This is a BREAKING CHANGE!

I think having a new-type for the digest is a good idea. We
can implement certain traits for it, like the `*Hex` traits
from `std::fmt`. We can also add methods (although I haven't
added any useful methods yet). So currently the type has the
following features:

- implements `LowerHex` and `UpperHex`
- implements `Index<usize>`
- implements `Deref<Target=[u8; 16]>`
- has method `raw_data()`, returning the internal array

It's questionable what features to add still. In particular,
whether we should make it behave like a slice. This means:
implementing all traits that are implemented for `[T]` or
rather `[T; 16]`. We *could* do it, but it would increase
the code size by quite a bit and maybe it's not useful. Maybe
it is, I don't know.

Anyway: the most common usage of the resulting digist is to
use the hexadecimal representation. This task shouldn't be
done by the user of this library, therefore I think the
new-type is required in order to provide a nice API. We might
even want to add a `hex_string()` method to provide a
shorthand for `format!("{:x}", digest)`.