Promote `Digest` from type alias to a real type (#3)
* Promote `Digest` from type alias to tuple struct

This is a BREAKING CHANGE!

I think having a new-type for the digest is a good idea. We
can implement certain traits for it, like the `*Hex` traits
from `std::fmt`. We can also add methods (although I haven't
added any useful methods yet). So currently the type has the
following features:

- implements `LowerHex` and `UpperHex`
- implements `Index<usize>`
- implements `Deref<Target=[u8; 16]>`
- has method `raw_data()`, returning the internal array

It's questionable what features to add still. In particular,
whether we should make it behave like a slice. This means:
implementing all traits that are implemented for `[T]` or
rather `[T; 16]`. We *could* do it, but it would increase
the code size by quite a bit and maybe it's not useful. Maybe
it is, I don't know.

Anyway: the most common usage of the resulting digist is to
use the hexadecimal representation. This task shouldn't be
done by the user of this library, therefore I think the
new-type is required in order to provide a nice API. We might
even want to add a `hex_string()` method to provide a
shorthand for `format!("{:x}", digest)`.

* Generalize `compute()` to accept everything `AsRef<[u8]>`

BREAKING CHANGE!
Although, it won't affect a whole lot of code, but only the code
which type inference relied on a fixed parameter type of `compute()`.

This makes the API more flexible and easy to easy. In particular,
it's possible to pass strings and string-slices directly to
`compute()`.

* Add documentation about `Context`

* Remove old code

* Cleanup, remove some documentation and add From impl

* Generalize `Context::consume` to AsRef<[u8]>

* Remove `raw_data()`